+++ babel.config.js
@@ -1,10 +1,10 @@
 module.exports = {
-  "presets": [
+  presets: [
     ["@vue/cli-plugin-babel/preset", 
       {
-      "useBuiltIns": false
+        useBuiltIns: false
       }
     ]
   ],
-  plugins: ['@babel/transform-runtime', 'babel-plugin-transform-import-meta'],
+  plugins: ["@babel/plugin-transform-runtime", "babel-plugin-transform-import-meta"],
 }